Added check for locales datagrid

+++ packages/Webkul/Admin/src/DataGrids/Settings/LocalesDataGrid.php
@@ -77,22 +77,26 @@ class LocalesDataGrid extends DataGrid
      */
     public function prepareActions()
     {
-        $this->addAction([
-            'icon'   => 'icon-edit',
-            'title'  => trans('admin::app.settings.locales.index.datagrid.edit'),
-            'method' => 'GET',
-            'url'    => function ($row) {
-                return route('admin.settings.locales.edit', $row->id);
-            },
-        ]);
+        if (bouncer()->hasPermission('settings.locales.edit')) {
+            $this->addAction([
+                'icon'   => 'icon-edit',
+                'title'  => trans('admin::app.settings.locales.index.datagrid.edit'),
+                'method' => 'GET',
+                'url'    => function ($row) {
+                    return route('admin.settings.locales.edit', $row->id);
+                },
+            ]);
+        }
 
-        $this->addAction([
-            'icon'   => 'icon-delete',
-            'title'  => trans('admin::app.settings.locales.index.datagrid.delete'),
-            'method' => 'DELETE',
-            'url'    => function ($row) {
-                return route('admin.settings.locales.delete', $row->id);
-            },
-        ]);
+        if (bouncer()->hasPermission('settings.locales.delete')) {
+            $this->addAction([
+                'icon'   => 'icon-delete',
+                'title'  => trans('admin::app.settings.locales.index.datagrid.delete'),
+                'method' => 'DELETE',
+                'url'    => function ($row) {
+                    return route('admin.settings.locales.delete', $row->id);
+                },
+            ]);
+        }
     }
 }
